+++ tests/python/pants_test/projects/test_python_testprojects_tests_integration.py
@@ -4,6 +4,6 @@
 from pants_test.projects.projects_test_base import ProjectsTestBase
 
 
-class TestPythonTestprojectsIntegration(ProjectsTestBase):
+class TestPythonTestprojectsTestsIntegration(ProjectsTestBase):
     def test_python_testprojects(self) -> None:
-        self.assert_valid_projects("testprojects/src/python::", "testprojects/tests/python::")
+        self.assert_valid_projects("testprojects/tests/python::")
